+++ ts/features/itwallet/issuance/screens/ItwIssuanceCredentialFailureScreen.tsx
@@ -28,7 +28,10 @@ import { ItwCredentialIssuanceMachineContext } from "../../machine/provider";
 import { useCredentialEventsTracking } from "../hooks/useCredentialEventsTracking";
 import { useIOSelector } from "../../../../store/hooks";
 import { itwDeferredIssuanceScreenContentSelector } from "../../../../store/reducers/backendStatus/remoteConfig";
-import { getFullLocale } from "../../../../utils/locale";
+import {
+  fallbackForLocalizedMessageKeys,
+  getFullLocale
+} from "../../../../utils/locale";
 import { serializeFailureReason } from "../../common/utils/itwStoreUtils";
 
 export const ItwIssuanceCredentialFailureScreen = () => {
@@ -65,6 +68,7 @@ const ContentView = ({ failure }: ContentViewProps) => {
     selectIssuerConfigurationOption
   );
   const locale = getFullLocale();
+  const localeFallback = fallbackForLocalizedMessageKeys(locale);
   const deferredIssuanceScreenContent = useIOSelector(
     itwDeferredIssuanceScreenContentSelector
   );
@@ -118,10 +122,10 @@ const ContentView = ({ failure }: ContentViewProps) => {
         case CredentialIssuanceFailureType.ASYNC_ISSUANCE:
           return {
             title:
-              deferredIssuanceScreenContent?.title?.[locale] ??
+              deferredIssuanceScreenContent?.title?.[localeFallback] ??
               I18n.t("features.itWallet.issuance.asyncCredentialError.title"),
             subtitle:
-              deferredIssuanceScreenContent?.description?.[locale] ??
+              deferredIssuanceScreenContent?.description?.[localeFallback] ??
               I18n.t("features.itWallet.issuance.asyncCredentialError.body"),
             pictogram: "pending",
             action: {
@@ -163,7 +167,12 @@ const ContentView = ({ failure }: ContentViewProps) => {
   });
 
   const resultScreenProps = getOperationResultScreenContentProps();
-  return <OperationResultScreenContent {...resultScreenProps} />;
+  return (
+    <OperationResultScreenContent
+      {...resultScreenProps}
+      subtitleProps={{ textBreakStrategy: "simple" }}
+    />
+  );
 };
 
 type GetCredentialInvalidStatusDetailsParams = {